// Frontend/jump-park-rh/src/types/apps/attendanceTypes.ts
export type AttendanceType = {
  id: number
  matricule: number
  date: string
  employee: string
  checkin: string
  checkout_pause: string
  checkin_pause: string
  checkout: string
  absence: string
  conge: string
  duration: string
  additional_hours: string
  createdAt: string
  status: string
}